+++ src/PickerField/i18n.js
@@ -1,0 +1,19 @@
+export default {
+  'zh-cn': {
+    selected: value => `已选择 ${value} 项`,
+    confirm: '确定',
+    complete: '完成',
+    noData: '无搜索结果',
+    searchPlaceholder: '请输入',
+    placeholder: '请选择',
+  },
+  'en-us': {
+    selected: value => `${value} items selected`,
+    confirm: 'OK',
+    complete: 'Complete',
+    noData: 'No Data',
+    searchPlaceholder: 'Please input',
+    placeholder: 'please select',
+  },
+};
+