+++ src/SearchBar/locale.js
@@ -1,10 +1,10 @@
 const locale = {
-  zh_CN: {
+  'zh-cn': {
     cancel: '取消',
     clear: '清空',
     history: '搜索历史',
   },
-  en_US: {
+  'en-us': {
     cancel: 'Cancel',
     clear: 'Clear',
     history: 'Search History',